Start to fix this animation craziness
I noticed that we had an issue where too many animations would try to
play at once when you switched apps. It appears we were in the middle of
a transition between using imperative animation running (the old way)
and declarative Transitions. This commit removes most of the imperative
animations to try to make sense of the situation.